refactor(mcp): move blankHiddenEnvValues into shared mcp utils
